requirements.txt
Pillow
pyautogui
tk
watchdog
google-cloud-vision
pyttsx3
keyboard
configparser